[jimmy] Step3 PR보냅니다. (#63)
* 로그인구현중, html header footer추가전

* redirect 에러 수정

* 로그인기능, 이슈추가 이슈삭제기능구현

* step3구현

+++ src/main/java/codesquad/service/IssueService.java
@@ -1,8 +1,11 @@
 package codesquad.service;
 
+import codesquad.CannotDeleteException;
 import codesquad.domain.Issue;
 import codesquad.domain.IssueRepository;
+import codesquad.domain.User;
 import codesquad.dto.IssueDto;
+import codesquad.security.LoginUser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -22,10 +25,9 @@ public class IssueService {
         return issueRepository.findAll();
     }
 
-    public Issue add(IssueDto issueDto) {
-        log.info("add method called");
+    public Issue add(User loginUser, IssueDto issueDto) {
         log.info("issueDto : {}", issueDto.toString());
-        return issueRepository.save(issueDto.toIssue());
+        return issueRepository.save(issueDto.toIssue().writeBy(loginUser));
     }
 
     public Issue findById(long id) {
@@ -34,15 +36,18 @@ public class IssueService {
     }
 
     @Transactional
-    public void update(long id, IssueDto updateIssue) {
+    public void update(User loginUser, long id, IssueDto updateIssue) throws CannotDeleteException {
         log.info("update method called");
         Issue original = issueRepository.findById(id).get();
-        original.update(updateIssue.toIssue());
+        original.update(loginUser, updateIssue.toIssue());
     }
 
     @Transactional
-    public void delete(long id) {
-        log.info("delete method called");
+    public void  delete(User loginUser, long id) throws CannotDeleteException {
+        Issue issue = issueRepository.findById(id).get();
+        if (!issue.isOwner(loginUser))
+            throw new CannotDeleteException("자신이 쓴 글만 삭제할 수 있습니다.");
+        log.info("delete success");
         issueRepository.deleteById(id);
     }
 }
